Remove Dependency on FST Serialization and roll my own.
This was to account for being able to query the database via the
database browser for remote databases.

Also, it is inline with the effort to remove 3rd party libraries

With migrating towards Android, this serialization will also take the
place of RMI when we remove the RMI dependency.

Also, I’ve added Proguard for jar optimization.